requirements.txt
Flask
twilio
openai
python-dotenv
gunicorn
google-generativeai
flask-cors
